requirements.txt
streamlit
langchain-upstage
langchain-core
python-dotenv
pandas